feat: complete immutability exercise
